Merge pull request #602 from jianchun:base
Remaining files depend on debugging, stack walker, or other unavailable
Windows APIs. To be address later.

Changes are:
- Added templates for a few InterLocked??? functions.
- Changed ```_NOEXCEPT``` to ```_NOEXCEPT_``` (was conflicting on VS 2013).
- "friend class ::ScriptMemoryDumper" needs "::" as the container classes
  are in a different "Memory" namespace.
- AuxPtr.h: added cast to use FieldEnum::Max as array size.
- Changed a __try/__finally to call TryFinally template function
- Standard ```__VA_ARGS__``` requires at least 1 arg. Found a well-known hack to
  work around: ```##__VA_ARGS__```. Verified works on VS 2013/2015, g++, clang.
- Added ENABLE_DEBUG_STACK_BACK_TRACE for capturing stack back traces for
  debugging only. (STACK_BACK_TRACE is defined on release.) Changed
  FunctionBody::cleanupReason under ENABLE_DEBUG_CONFIG_OPTIONS and
  cleanupStack under ENABLE_DEBUG_STACK_BACK_TRACE.

+++ lib/Runtime/Base/AuxPtrs.h
@@ -7,7 +7,7 @@
 
 namespace Js
 {
-    // Use fixed size structure to save pointers 
+    // Use fixed size structure to save pointers
     // AuxPtrsFix(16 bytes or 32 bytes) layout:
     //                        max count  metadata(init'd type)        pointers array
     //     --------------------------------------------------------------------------------------------------
@@ -28,7 +28,7 @@ namespace Js
         bool Set(FieldsEnum e, void* p);
     };
 
-    // Use flexible size structure to save pointers. when pointer count exceeds AuxPtrsFix<FieldsEnum, 32>::MaxCount, 
+    // Use flexible size structure to save pointers. when pointer count exceeds AuxPtrsFix<FieldsEnum, 32>::MaxCount,
     // it will promote to this structure to save the pointers
     // Layout:
     //     count      array of positions       array of instantiated pointers
@@ -42,7 +42,7 @@ namespace Js
         typedef AuxPtrs<T, FieldsEnum> AuxPtrsT;
         uint8 count;                            // save instantiated pointers count
         uint8 capacity;                         // save number of pointers can be hold in current instance of AuxPtrs
-        uint8 offsets[FieldsEnum::Max];       // save position of each instantiated pointers, if not instantiate, it's invalid
+        uint8 offsets[static_cast<int>(FieldsEnum::Max)];       // save position of each instantiated pointers, if not instantiate, it's invalid
         WriteBarrierPtr<void> ptrs[1];          // instantiated pointer addresses
         AuxPtrs(uint8 capacity, AuxPtrs32* ptr32);  // called when promoting from AuxPtrs32 to AuxPtrs
         AuxPtrs(uint8 capacity, AuxPtrs* ptr);      // called when expanding (i.e. promoting from AuxPtrs to bigger AuxPtrs)
